+++ Minecraft.World/BrewingStandTileEntity.cpp
@@ -305,7 +305,7 @@ void BrewingStandTileEntity::save(CompoundTag *base)
         if (items[i] != NULL)
         {
             CompoundTag *tag = new CompoundTag();
-            tag->putByte(L"Slot", (byte)i);
+            tag->putByte(L"Slot", i);
             items[i]->save(tag);
             listTag->add(tag);
         }
