+++ Minecraft.World/Creeper.cpp
@@ -61,8 +61,8 @@ void Creeper::defineSynchedData()
 {
     Monster::defineSynchedData();
 
-    entityData->define(DATA_SWELL_DIR, (byte)-1);
-    entityData->define(DATA_IS_POWERED, (byte)0);
+    entityData->define(DATA_SWELL_DIR, -1);
+    entityData->define(DATA_IS_POWERED, 0);
 }
 
 void Creeper::addAdditonalSaveData(CompoundTag *entityTag)
@@ -77,7 +77,7 @@ void Creeper::addAdditonalSaveData(CompoundTag *entityTag)
 void Creeper::readAdditionalSaveData(CompoundTag *tag)
 {
     Monster::readAdditionalSaveData(tag);
-    entityData->set(DATA_IS_POWERED, (byte)(tag->getBoolean(L"powered") ? 1 : 0));
+    entityData->set(DATA_IS_POWERED, (tag->getBoolean(L"powered") ? 1 : 0));
 }
 
 void Creeper::tick()
@@ -169,11 +169,11 @@ int Creeper::getSwellDir()
 
 void Creeper::setSwellDir(int dir)
 {
-    entityData->set(DATA_SWELL_DIR, (byte)dir);
+    entityData->set(DATA_SWELL_DIR, dir);
 }
 
 void Creeper::thunderHit(const LightningBolt *lightningBolt)
 {
     Monster::thunderHit(lightningBolt);
-    entityData->set(DATA_IS_POWERED, (byte)1);
+    entityData->set(DATA_IS_POWERED, 1);
 }
