+++ Minecraft.World/EnderMan.cpp
@@ -62,9 +62,9 @@ void EnderMan::defineSynchedData()
 {
     Monster::defineSynchedData();
 
-    entityData->define(DATA_CARRY_ITEM_ID, (byte)0);
-    entityData->define(DATA_CARRY_ITEM_DATA, (byte)0);
-    entityData->define(DATA_CREEPY, (byte)0);
+    entityData->define(DATA_CARRY_ITEM_ID, 0);
+    entityData->define(DATA_CARRY_ITEM_DATA, 0);
+    entityData->define(DATA_CREEPY, 0);
 }
 
 void EnderMan::addAdditonalSaveData(CompoundTag *tag)
@@ -374,7 +374,7 @@ void EnderMan::dropDeathLoot(bool wasKilledByPlayer, int playerBonusLevel)
 // 4J Brought forward from 1.2.3 to help fix Enderman behaviour
 void EnderMan::setCarryingTile(int carryingTile)
 {
-    entityData->set(DATA_CARRY_ITEM_ID, (byte)(carryingTile & 0xff));
+    entityData->set(DATA_CARRY_ITEM_ID, (carryingTile & 0xff));
 }
 
 int EnderMan::getCarryingTile()
@@ -384,7 +384,7 @@ int EnderMan::getCarryingTile()
 
 void EnderMan::setCarryingData(int carryingData)
 {
-    entityData->set(DATA_CARRY_ITEM_DATA, (byte)(carryingData & 0xff));
+    entityData->set(DATA_CARRY_ITEM_DATA, (carryingData & 0xff));
 }
 
 int EnderMan::getCarryingData()
@@ -415,5 +415,5 @@ bool EnderMan::isCreepy()
 
 void EnderMan::setCreepy(bool creepy)
 {
-    entityData->set(DATA_CREEPY, (byte)(creepy ? 1 : 0));
+    entityData->set(DATA_CREEPY, (creepy ? 1 : 0));
 }
