+++ Minecraft.World/MusicTileEntity.cpp
@@ -37,7 +37,7 @@ void MusicTileEntity::load(CompoundTag *tag)
 
 void MusicTileEntity::tune()
 {
-    note = (byte)((note + 1) % 25);
+    note = ((note + 1) % 25);
     setChanged();
 }
