+++ Minecraft.World/TileEntityDataPacket.cpp
@@ -47,7 +47,7 @@ void TileEntityDataPacket::write(DataOutputStream *dos)
     dos->writeInt(x);
     dos->writeShort(y);
     dos->writeInt(z);
-    dos->writeByte((byte)type);
+    dos->writeByte(type);
     writeNbt(tag, dos);
 }
